pyproject.toml
[tool.poetry]
name = "tajana"
version = "0.1.0"
description = "test files"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "data_pipelines" },
    { include = "src" },
    { include = "research_experiments" }
]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
vectorbt = "^0.26.2"
yfinance = "^0.2.41"
nbformat = "^5.10.4"
xbbg = "^0.7.7"
quantstats = "^0.0.62"
ipykernel = "^6.29.5"
plotly = "^5.23.0"
openpyxl = "^3.1.5"
jinja2 = "^3.1.4"
setuptools = "^73.0.1"
dash = "^2.17.1"
jupyter-dash = "^0.4.2"
grid = "^0.7.1"
qgrid = "^1.3.1"
chart-studio = "^1.1.0"
nbconvert = {extras = ["webpdf"], version = "^7.16.4"}
notebook = "^7.2.2"
playwright = "^1.47.0"
sqlalchemy = "^2.0.25"
pyyaml = "^6.0.1"
jupyter = "^1.1.1"
pdblp = "^0.1.8"
seaborn = "^0.13.2"
dash-bootstrap-components = "^1.6.0"
requests = "^2.32.3"
ipywidgets = "^8.1.5"
weasyprint = "^63.1"
lightgbm = "^4.5.0"
xgboost = "^2.1.3"
optuna = "^4.1.0"
filterpy = "^1.4.5"
hmmlearn = "^0.3.0"
flaml = "^2.3.3"
autots = "^0.6.17"
scikit-learn = "^1.6.1"
pandas = "<2.0.0"
numpy = "<2.0.0"
arch = "^7.2.0"
torch = "^2.5.1"
deap = "^1.4.2"
ruptures = "^1.1.9"
catboost = "^1.2.7"
tsfresh = "^0.21.0"
fredapi = "^0.5.2"
matplotlib = "^3.10.1"
supabase = "^2.15.0"
python-dotenv = "^1.1.0"
rapidfuzz = "^3.13.0"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
